// server/types.ts
export type UserRole = 'Recruiter' | 'Admin';

export interface User {
  id: string;
  name: string;
  email: string;
  phone: string;
  company: string;
  passwordHash: string;
  role: UserRole;
  status?: 'Active' | 'Disabled';
  createdAt: string;
  updatedAt: string;
}